express api, sequelize model, and seed file working 6/6.  All specs for Tier 1 Robot working 18/18 with custom tests added.

+++ server/db/robot.js
@@ -38,7 +38,9 @@ const Robot = db.define("robot", {
     defaultValue: "electric",
   },
   fuelLevel: {
-    type: Sequelize.DECIMAL(10, 2),
+    // !REMOVE - had to use FLOAT to pass the specs
+    // type: Sequelize.DECIMAL(10, 2),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true,
